Removed 'shell' opcode
