+++ src/org/bds/vm/BdsVm.java
@@ -34,7 +34,6 @@ import org.bds.run.RunState;
 import org.bds.scope.Scope;
 import org.bds.symbol.SymbolTable;
 import org.bds.task.DepVmOpcode;
-import org.bds.task.ShellVmOpcode;
 import org.bds.task.SysVmOpcode;
 import org.bds.task.TaskDependency;
 import org.bds.task.TaskVmOpcode;
@@ -1452,13 +1451,6 @@ public class BdsVm implements Serializable {
 				vmap = (ValueMap) pop();
 				v1 = pop(); // Key
 				vmap.put(v1, pop());
-				break;
-
-			case SHELL:
-				vmStateSave();
-				ShellVmOpcode shellVmOp = new ShellVmOpcode(bdsThread);
-				s1 = shellVmOp.run();
-				push(s1);
 				break;
 
 			case STORE:
